sspi: move AuthIdentity and AuthIdentityBuffers into a separate module
