+++ src/credssp/mod.rs
@@ -24,14 +24,15 @@ use crate::builders::{ChangePassword, EmptyInitializeSecurityContext};
 use crate::crypto::compute_sha256;
 use crate::kerberos::config::KerberosConfig;
 use crate::kerberos::{self, Kerberos};
-use crate::ntlm::{self, AuthIdentity, AuthIdentityBuffers, Ntlm, NtlmConfig, SIGNATURE_SIZE};
+use crate::ntlm::{self, Ntlm, SIGNATURE_SIZE, NtlmConfig};
 use crate::pku2u::{self, Pku2u, Pku2uConfig};
 use crate::{
-    negotiate, AcceptSecurityContextResult, AcquireCredentialsHandleResult, CertContext, CertTrustStatus,
-    ClientRequestFlags, ConnectionInfo, ContextNames, ContextSizes, CredentialUse, DataRepresentation, DecryptionFlags,
-    EncryptionFlags, Error, ErrorKind, FilledAcceptSecurityContext, FilledAcquireCredentialsHandle,
-    FilledInitializeSecurityContext, InitializeSecurityContextResult, Negotiate, NegotiateConfig, PackageInfo,
-    SecurityBuffer, SecurityBufferType, SecurityStatus, ServerRequestFlags, Sspi, SspiEx, SspiImpl, StreamSizes,
+    negotiate, AcceptSecurityContextResult, AcquireCredentialsHandleResult, AuthIdentity, AuthIdentityBuffers,
+    CertContext, CertTrustStatus, ClientRequestFlags, ConnectionInfo, ContextNames, ContextSizes, CredentialUse,
+    DataRepresentation, DecryptionFlags, EncryptionFlags, Error, ErrorKind, FilledAcceptSecurityContext,
+    FilledAcquireCredentialsHandle, FilledInitializeSecurityContext, InitializeSecurityContextResult, Negotiate,
+    NegotiateConfig, PackageInfo, SecurityBuffer, SecurityBufferType, SecurityStatus, ServerRequestFlags, Sspi, SspiEx,
+    SspiImpl, StreamSizes,
 };
 
 pub const EARLY_USER_AUTH_RESULT_PDU_SIZE: usize = 4;
