+++ src/credssp/ts_request/test.rs
@@ -2,8 +2,7 @@ use lazy_static::lazy_static;
 
 use super::*;
 use crate::credssp::CredSspMode;
-use crate::ntlm::AuthIdentityBuffers;
-use crate::AuthIdentity;
+use crate::{AuthIdentity, AuthIdentityBuffers};
 
 const NTLM_CLIENT_NONCE: [u8; 32] = [
     0x22, 0x10, 0x12, 0xad, 0x12, 0x5c, 0x7a, 0x15, 0xfe, 0xb6, 0x4b, 0x1f, 0xcb, 0x94, 0x83, 0x3a, 0xc5, 0x6f, 0x66,
