+++ src/ntlm/mod.rs
@@ -10,17 +10,16 @@ use bitflags::bitflags;
 use byteorder::{LittleEndian, WriteBytesExt};
 use lazy_static::lazy_static;
 use messages::{client, server};
-use serde_derive::{Deserialize, Serialize};
 
 pub use self::config::NtlmConfig;
 use super::channel_bindings::ChannelBindings;
 use crate::crypto::{compute_hmac_md5, Rc4, HASH_SIZE};
 use crate::{
-    utils, AcceptSecurityContextResult, AcquireCredentialsHandleResult, CertTrustStatus, ClientResponseFlags,
-    ContextNames, ContextSizes, CredentialUse, DecryptionFlags, EncryptionFlags, FilledAcceptSecurityContext,
-    FilledAcquireCredentialsHandle, FilledInitializeSecurityContext, InitializeSecurityContextResult,
-    PackageCapabilities, PackageInfo, Result, Secret, SecurityBuffer, SecurityBufferType, SecurityPackageType,
-    SecurityStatus, ServerResponseFlags, Sspi, SspiEx, SspiImpl, PACKAGE_ID_NONE,
+    AcceptSecurityContextResult, AcquireCredentialsHandleResult, AuthIdentity, AuthIdentityBuffers, CertTrustStatus,
+    ClientResponseFlags, ContextNames, ContextSizes, CredentialUse, DecryptionFlags, EncryptionFlags,
+    FilledAcceptSecurityContext, FilledAcquireCredentialsHandle, FilledInitializeSecurityContext,
+    InitializeSecurityContextResult, PackageCapabilities, PackageInfo, Result, SecurityBuffer, SecurityBufferType,
+    SecurityPackageType, SecurityStatus, ServerResponseFlags, Sspi, SspiEx, SspiImpl, PACKAGE_ID_NONE,
 };
 
 pub const PKG_NAME: &str = "NTLM";
@@ -500,95 +499,6 @@ impl AuthenticateMessage {
 impl Mic {
     fn new(value: [u8; MESSAGE_INTEGRITY_CHECK_SIZE], offset: u8) -> Self {
         Self { value, offset }
-    }
-}
-
-/// Allows you to pass a particular user name and password to the run-time library for the purpose of authentication
-///
-/// # MSDN
-///
-/// * [SEC_WINNT_AUTH_IDENTITY_W structure](https://docs.microsoft.com/en-us/windows/win32/api/sspi/ns-sspi-sec_winnt_auth_identity_w)
-#[derive(Debug, Clone, Eq, PartialEq, Default, Serialize, Deserialize)]
-pub struct AuthIdentity {
-    pub username: String,
-    pub password: Secret<String>,
-    pub domain: Option<String>,
-}
-
-#[derive(Clone, Eq, PartialEq, Default)]
-pub struct AuthIdentityBuffers {
-    pub user: Vec<u8>,
-    pub domain: Vec<u8>,
-    pub password: Secret<Vec<u8>>,
-}
-
-impl AuthIdentityBuffers {
-    pub fn new(user: Vec<u8>, domain: Vec<u8>, password: Vec<u8>) -> Self {
-        Self {
-            user,
-            domain,
-            password: Secret::new(password),
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.user.is_empty()
-    }
-}
-
-impl Debug for AuthIdentityBuffers {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "AuthIdentityBuffers {{ user: 0x")?;
-        self.user.iter().try_for_each(|byte| write!(f, "{byte:02X}"))?;
-        write!(f, ", domain: 0x")?;
-        self.domain.iter().try_for_each(|byte| write!(f, "{byte:02X}"))?;
-        write!(f, ", password: {:?} }}", self.password)?;
-
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{AuthIdentityBuffers, Secret};
-
-    #[test]
-    fn td() {
-        let auth = AuthIdentityBuffers {
-            user: vec![98, 0, 50, 0],
-            domain: vec![69, 0, 88, 0, 65, 0, 77, 0, 80, 0, 76, 0, 69, 0],
-            password: Secret::new(vec![
-                113, 0, 113, 0, 113, 0, 81, 0, 81, 0, 81, 0, 49, 0, 49, 0, 49, 0, 33, 0, 33, 0, 33, 0,
-            ]),
-        };
-        println!("{:?}", auth);
-    }
-}
-
-impl From<AuthIdentity> for AuthIdentityBuffers {
-    fn from(credentials: AuthIdentity) -> Self {
-        Self {
-            user: utils::string_to_utf16(credentials.username.as_str()),
-            domain: credentials
-                .domain
-                .map(|v| utils::string_to_utf16(v.as_str()))
-                .unwrap_or_default(),
-            password: utils::string_to_utf16(credentials.password.as_ref()).into(),
-        }
-    }
-}
-
-impl From<AuthIdentityBuffers> for AuthIdentity {
-    fn from(credentials_buffers: AuthIdentityBuffers) -> Self {
-        Self {
-            username: utils::bytes_to_utf16_string(credentials_buffers.user.as_ref()),
-            password: utils::bytes_to_utf16_string(credentials_buffers.password.as_ref()).into(),
-            domain: if credentials_buffers.domain.is_empty() {
-                None
-            } else {
-                Some(utils::bytes_to_utf16_string(credentials_buffers.domain.as_ref()))
-            },
-        }
     }
 }
